refactor: UserMapper와 UniversityMapper 분리 (#314)
