+++ server/momo-core/src/main/java/com/momo/user/application/UniversityServiceImpl.java
@@ -1,6 +1,6 @@
 package com.momo.user.application;
 
-import com.momo.user.application.dto.UserDtoMapper;
+import com.momo.user.application.dto.UniversityDtoMapper;
 import com.momo.user.application.dto.response.UniversityResponseDto;
 import com.momo.user.domain.UniversitySearchOpenApi;
 import com.momo.user.domain.dto.UniversityDto;
@@ -13,11 +13,11 @@ import org.springframework.stereotype.Service;
 public class UniversityServiceImpl implements UniversityService {
 
     private final UniversitySearchOpenApi universitySearchOpenApi;
-    private final UserDtoMapper userDtoMapper;
+    private final UniversityDtoMapper universityDtoMapper;
 
     public List<UniversityResponseDto> findUniversity(String universityName) {
         List<UniversityDto> universityDtos = universitySearchOpenApi.search(universityName);
 
-        return userDtoMapper.mapToUniversityResponseDtos(universityDtos);
+        return universityDtoMapper.mapToUniversityResponseDtos(universityDtos);
     }
 }
